Alphabetize channels and sort vids by publish date
Channels are now alphabetical and videos are sorted from most recently published to least recently. Some problem occurred with at least one channel (Mr. Sark) where the 5 videos fetched weren't the 5 most recent videos on the channel. There's probably a problem with publish dates. Before, some resources fetched from the api list the publish date as the date the owner uploaded it to youtube rather than the date it was visible to the public. This comes up when an owner uploads in advance such as before going on vacation.